computing principal stress in double
