+++ pyansys/rst.py
@@ -1396,11 +1396,6 @@ class ResultFile(AnsysBinary):
         """
         # get component stress
         nodenum, stress = self.nodal_stress(rnum)
-
-        # compute principle stress
-        if stress.dtype != np.float32:
-            stress = stress.astype(np.float32)
-
         pstress, isnan = _binary_reader.compute_principal_stress(stress)
         pstress[isnan] = np.nan
         return nodenum, pstress
@@ -1954,12 +1949,12 @@ class ResultFile(AnsysBinary):
         # Element types for nodal averaging
         elemtype = self.geometry['Element Type'].astype(np.int32)
 
-        if self.version < 14.5:  # values stored as double precision
-            tarr = np.empty(1, np.float64)
-            my_dtype = 1
-        else:    # values stored as single precision
-            tarr = np.empty(1, np.float32)
-            my_dtype = 0
+        # if self.version < 14.5:  # values stored as double precision
+        # tarr = np.empty(1, np.float64)
+            # my_dtype = 1
+        # else:    # values stored as single precision
+        #     tarr = np.empty(1, np.float32)
+        #     my_dtype = 0
 
         data, ncount = _binary_reader.read_nodal_values(self.filename,
                                                         self.grid.celltypes,
@@ -1971,9 +1966,7 @@ class ResultFile(AnsysBinary):
                                                         nodstr,
                                                         etype,
                                                         elemtype,
-                                                        result_index,
-                                                        tarr,
-                                                        my_dtype)
+                                                        result_index)
 
         if result_type == 'ENS' and nitem != 6:
             data = data[:, :6]
