+++ pytest.ini
@@ -3,3 +3,6 @@ filterwarnings =
     ignore::FutureWarning
     ignore::PendingDeprecationWarning
     ignore::DeprecationWarning
+    # bogus numpy ABI warning (see numpy/#432)
+    ignore:.*numpy.dtype size changed.*:RuntimeWarning
+    ignore:.*numpy.ufunc size changed.*:RuntimeWarning
